Moved nldi legend into the legend module and deduplicated some now-shared logic
